+++ tests/unit/test_sonata/test_split_population.py
@@ -1,4 +1,3 @@
-import itertools as it
 import os
 from pathlib import Path