+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
   * Add simple circuit splitting for SONATA (BRBLD-90).
   * Add subcircuit splitting for SONATA (BRBLD-90).
   * teach `node-set-from-targets` to take paths to files, add to documentation
+  * Add support for copying the used morphologies from a SONATA circuit (BRBLD-93).
 
 ## 0.17.0
   * Update ``write_network_config`` to be compatible with the new BBP sonata specs, moved from